error formatting based on documentation

+++ server/src/api/v1/article/controllers/deleteItem.js
@@ -1,8 +1,18 @@
 const articleServices = require("../../../../lib/articles");
+const { badRequest } = require("../../../../utils/error");
 
 const deleteItem = async (req, res, next) => {
-  const { id } = req.params;
   try {
+    const { id } = req.params;
+
+    // Validate id
+    if (!id || typeof id !== "string") {
+      throw badRequest(
+        [{ field: "id", message: "invalid input", in: "params" }],
+        "invalid input",
+      );
+    }
+
     await articleServices.deleteItem(id);
     res.status(204).end();
   } catch (e) {
